requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
aiocache==0.12.3
python-dotenv==1.0.1
pytest==8.3.0
ruff==0.6.0
pytest-asyncio==0.24.0